backend/requirements.txt
Flask>=2.0
Flask-CORS>=4.0.0
google-generativeai>=0.3.0
cryptography>=41.0.0
python-dotenv>=1.0.0
requests>=2.31.0
chromadb>=0.4.0
sentence-transformers>=2.2.0